Refactor AG10 send identity helpers
