+++ crates/atm-core/src/send/nudge_template.rs
@@ -1,11 +1,12 @@
 use std::collections::BTreeMap;
 
-use crate::address::AgentAddress;
 use crate::boundary::{
     BuiltInNudgeTemplateKind, PostSendHookEvent, ResolvedBuiltInNudgeTemplate,
     TeamNudgeTemplateOverrideRow,
 };
 use crate::error::AtmError;
+
+use super::qualified_sender_identity;
 
 pub fn resolve_template(
     override_row: Option<TeamNudgeTemplateOverrideRow>,
@@ -23,14 +24,6 @@ pub fn resolve_template_body(
     kind: BuiltInNudgeTemplateKind,
 ) -> Option<String> {
     resolve_template(override_row, kind).body
-}
-
-pub fn qualified_sender_identity(event: &PostSendHookEvent) -> String {
-    AgentAddress {
-        agent: event.sender.clone(),
-        team: Some(event.sender_team.clone()),
-    }
-    .to_string()
 }
 
 pub fn render_resolved_built_in_nudge(
@@ -59,7 +52,10 @@ pub fn render_built_in_nudge(
 
 fn render_values(event: &PostSendHookEvent) -> BTreeMap<&'static str, String> {
     BTreeMap::from([
-        ("from", qualified_sender_identity(event)),
+        (
+            "from",
+            qualified_sender_identity(&event.sender, Some(&event.sender_team)),
+        ),
         ("team", event.recipient_team.to_string()),
         ("message_id", event.message_id.to_string()),
         ("description", event.description.clone()),
@@ -139,14 +135,12 @@ fn render_template(
 
 #[cfg(test)]
 mod tests {
-    use super::{
-        default_template, qualified_sender_identity, render_built_in_nudge, resolve_template,
-        resolve_template_body,
-    };
+    use super::{default_template, render_built_in_nudge, resolve_template, resolve_template_body};
     use crate::boundary::{
         BuiltInNudgeTemplateKind, PostSendHookEvent, ResolvedBuiltInNudgeTemplate,
         TeamNudgeTemplateOverrideMode, TeamNudgeTemplateOverrideRow,
     };
+    use crate::send::qualified_sender_identity;
     use crate::test_support::{TEST_ARCH_CTM, TEST_LEAD, TEST_TEAM};
     use crate::types::{AgentName, IsoTimestamp, PaneId, TeamName};
 
@@ -209,7 +203,7 @@ mod tests {
     #[test]
     fn qualified_sender_identity_uses_sender_and_team() {
         assert_eq!(
-            qualified_sender_identity(&base_event()),
+            qualified_sender_identity(&base_event().sender, Some(&base_event().sender_team),),
             format!("{TEST_LEAD}@{TEST_TEAM}")
         );
     }
