+++ crates/atm-core/src/send/warning.rs
@@ -1,0 +1,46 @@
+use crate::error_codes::AtmErrorCode;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct WarningEntry {
+    pub message: String,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub code: Option<AtmErrorCode>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub recovery: Option<String>,
+}
+
+impl WarningEntry {
+    pub fn new(message: impl Into<String>, recovery: Option<impl Into<String>>) -> Self {
+        Self {
+            message: message.into(),
+            code: None,
+            recovery: recovery.map(Into::into),
+        }
+    }
+
+    pub fn with_code(
+        code: AtmErrorCode,
+        message: impl Into<String>,
+        recovery: Option<impl Into<String>>,
+    ) -> Self {
+        Self {
+            message: message.into(),
+            code: Some(code),
+            recovery: recovery.map(Into::into),
+        }
+    }
+
+    pub fn render(&self) -> String {
+        let message = match self.code {
+            Some(code) if !self.message.contains(code.as_str()) => {
+                format!("{} [{}]", self.message, code.as_str())
+            }
+            _ => self.message.clone(),
+        };
+        match &self.recovery {
+            Some(recovery) => format!("{message} Recovery: {recovery}"),
+            None => message,
+        }
+    }
+}
